added discriminator key to user schema, reated buyer schema extension
