+++ server/db/models/user.js
@@ -2,7 +2,6 @@
 var crypto = require('crypto');
 var mongoose = require('mongoose');
 var Schema = mongoose.Schema;
-var extend = require('mongoose-schema-extend');
 var _ = require('lodash');
 
 var userSchema = new Schema({
@@ -41,25 +40,10 @@ var userSchema = new Schema({
     },
     picture: String,
     transactions: [{type: Schema.Types.ObjectId, ref: 'Transaction'}]
-});
-
-var chefSchema = userSchema.extend({
-  specialty: String,
-  bio: String,
-  rating: {
-    type: Number,
-    default: 0
-  },
-  meals: {
-    type: Schema.Types.ObjectId, 
-    ref: 'Meals'
-  }
-});
-
-
+}, {collection: 'users', discriminatorKey: '_type'});
 
 // method to remove sensitive information from user objects before sending them out
-schema.methods.sanitize =  function () {
+userSchema.methods.sanitize =  function () {
     return _.omit(this.toJSON(), ['password', 'salt']);
 };
 
@@ -76,7 +60,7 @@ var encryptPassword = function (plainText, salt) {
     return hash.digest('hex');
 };
 
-schema.pre('save', function (next) {
+userSchema.pre('save', function (next) {
 
     if (this.isModified('password')) {
         this.salt = this.constructor.generateSalt();
@@ -87,11 +71,11 @@ schema.pre('save', function (next) {
 
 });
 
-schema.statics.generateSalt = generateSalt;
-schema.statics.encryptPassword = encryptPassword;
+userSchema.statics.generateSalt = generateSalt;
+userSchema.statics.encryptPassword = encryptPassword;
 
-schema.method('correctPassword', function (candidatePassword) {
+userSchema.method('correctPassword', function (candidatePassword) {
     return encryptPassword(candidatePassword, this.salt) === this.password;
 });
 
-mongoose.model('User', schema);
+mongoose.model('User', userSchema);
